+++ IBS.DataAccess/Repository/Msap/IRepository/IDispatchTicketRepository.cs
@@ -1,6 +1,7 @@
 using IBS.DataAccess.Repository.IRepository;
 using IBS.Models;
 using IBS.Models.MSAP;
+using IBS.Models.MSAP.ViewModels;
 
 namespace IBS.DataAccess.Repository.Msap.IRepository
 {
@@ -15,6 +16,8 @@ namespace IBS.DataAccess.Repository.Msap.IRepository
         Task<IEnumerable<DispatchTicket>> GetDispatchTicketsWithDetailsAsync(DateTime start, DateTime end, CancellationToken cancellationToken = default);
 
         Task<bool> IsJobOrderEditableAsync(int? jobOrderId, CancellationToken cancellationToken = default);
+
+        Task<DispatchTicketViewModel> GetDispatchTicketSelectLists(DispatchTicketViewModel model, CancellationToken cancellationToken = default);
 
         Task<(IEnumerable<DispatchTicket> Data, int RecordsFiltered, int TotalRecords)> GetPagedDispatchTicketsAsync(DataTablesParameters parameters, string filterType, CancellationToken cancellationToken = default);       }
 }