+++ IBS.Models/ViewModels/DashboardCountViewModel.cs
@@ -2,7 +2,6 @@ namespace IBS.Models.ViewModels
 {
     public class DashboardCountViewModel
     {
-        public int MsapServiceRequestForPosting { get; set; }
         public int MsapDispatchTicketForTariff { get; set; }
         public int MsapDispatchTicketForApproval { get; set; }
         public int MsapDispatchTicketForBilling { get; set; }